Refactoring codegenerator.
Classes that might be shared by other generator has been moved to
quakus-xapi-generator-common dependepency.
