+++ deployment/pom.xml
@@ -13,6 +13,7 @@
     <version.org.openapitools>5.4.0</version.org.openapitools>
     <version.org.slf4j>2.0.3</version.org.slf4j>
     <version.com.github.jknack>4.3.1</version.com.github.jknack>
+    <io.quarkiverse.xapi.generator>1.0.0</io.quarkiverse.xapi.generator>
   </properties>
 
   <dependencies>
@@ -36,8 +37,13 @@
     <dependency>
       <groupId>io.quarkiverse.openapi.generator</groupId>
       <artifactId>quarkus-openapi-generator</artifactId>
-      <version>${project.version}</version>
     </dependency>
+   <dependency>
+     <groupId>io.quarkiverse.xapi.generator</groupId>
+     <artifactId>quarkus-xapi-generator-common</artifactId>
+     <version>${io.quarkiverse.xapi.generator}</version>
+   </dependency>
+
 
     <!-- OpenApi Generator -->
     <dependency>
