+++ deployment/src/main/java/io/quarkiverse/openapi/generator/deployment/codegen/OpenApiGeneratorJsonCodeGen.java
@@ -1,12 +1,6 @@
 package io.quarkiverse.openapi.generator.deployment.codegen;
 
 public class OpenApiGeneratorJsonCodeGen extends OpenApiGeneratorCodeGenBase {
-
-    @Override
-    public String providerId() {
-        return OpenApiGeneratorOutputPaths.JSON_PATH;
-    }
-
     @Override
     public String inputExtension() {
         return JSON;
