+++ deployment/src/main/java/io/quarkiverse/openapi/generator/deployment/codegen/OpenApiGeneratorYmlCodeGen.java
@@ -3,11 +3,6 @@ package io.quarkiverse.openapi.generator.deployment.codegen;
 public class OpenApiGeneratorYmlCodeGen extends OpenApiGeneratorCodeGenBase {
 
     @Override
-    public String providerId() {
-        return OpenApiGeneratorOutputPaths.YML_PATH;
-    }
-
-    @Override
     public String inputExtension() {
         return YML;
     }
